Renamed and unify fixedPoint
There where two different implementations of the same thing:
iterateUntilConverged and the repeat rewriter. This commit throws one away,
 and renames the remaining one to fixedPoint, the mathematical name for
 this concept.
