+++ community/cypher/cypher-compiler-3.0/src/main/scala/org/neo4j/cypher/internal/compiler/v3_0/ast/rewriters/inlineProjections.scala
@@ -20,8 +20,8 @@
 package org.neo4j.cypher.internal.compiler.v3_0.ast.rewriters
 
 import org.neo4j.cypher.internal.frontend.v3_0.ast._
-import org.neo4j.cypher.internal.compiler.v3_0.helpers.Converge.iterateUntilConverged
 import org.neo4j.cypher.internal.compiler.v3_0.planner.CantHandleQueryException
+import org.neo4j.cypher.internal.frontend.v3_0.helpers.fixedPoint
 import org.neo4j.cypher.internal.frontend.v3_0.{replace, Rewriter, TypedRewriter}
 
 case object inlineProjections extends Rewriter {
@@ -73,7 +73,7 @@ case object inlineProjections extends Rewriter {
   }
 
   private def findAllDependencies(variable: Variable, context: InliningContext): Set[Variable] = {
-    val (dependencies, _) = iterateUntilConverged[(Set[Variable], List[Variable])]({
+    val (dependencies, _) = fixedPoint[(Set[Variable], List[Variable])]({
       case (deps, Nil) =>
         (deps, Nil)
       case (deps, queue) =>
