+++ community/cypher/cypher-compiler-3.0/src/test/scala/org/neo4j/cypher/internal/compiler/v3_0/planner/logical/plans/rewriter/UnnestEmptyApplyTest.scala
@@ -19,10 +19,10 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_0.planner.logical.plans.rewriter
 
-import org.neo4j.cypher.internal.compiler.v3_0.helpers.Converge.iterateUntilConverged
 import org.neo4j.cypher.internal.compiler.v3_0.planner.LogicalPlanningTestSupport
 import org.neo4j.cypher.internal.compiler.v3_0.planner.logical.plans._
 import org.neo4j.cypher.internal.frontend.v3_0.SemanticDirection
+import org.neo4j.cypher.internal.frontend.v3_0.helpers.fixedPoint
 import org.neo4j.cypher.internal.frontend.v3_0.test_helpers.CypherFunSuite
 
 class UnnestEmptyApplyTest extends CypherFunSuite with LogicalPlanningTestSupport {
@@ -232,5 +232,5 @@ class UnnestEmptyApplyTest extends CypherFunSuite with LogicalPlanningTestSuppor
   }
 
   private def rewrite(p: LogicalPlan): LogicalPlan =
-    iterateUntilConverged((p: LogicalPlan) => p.endoRewrite(unnestApply))(p)
+    fixedPoint((p: LogicalPlan) => p.endoRewrite(unnestApply))(p)
 }
